Support building local docker (#1512)
Right now, the docker image we build as a part of CI for skypilot isn't
compatible with mac (arch issue). Coercing it to be without changing the
base pufferlib image was hard

This builds a docker image that avoids that, and as a bonus, uses
`./install.sh` to install what it needs (other than brew) -- pretty
exciting!

Also adds `metta local {x}` for us to put local development commands in.
I would like to put all the commands for spinning up local observatory
in there, for instance. For now it has `metta local build-docker-img`

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210816885719034)

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>