Delete fast_gae (#1498)
Remove our version of fast_gae

We use pufferlib to compute advantage. In particular, our advantage
calculation is done via `torch.ops.pufferlib.compute_puff_advantage`.


[`compute_puff_advantage`](https://github.com/PufferAI/PufferLib/blob/600c55aa03c767ee2010fcd1a11f38b1eaef7ae8/pufferlib/extensions/pufferlib.cpp#L75)
looks equivalent to fast_gae, except
* it has a cuda implementation, as well as cpu
* it also considers importance, so it can presumably be used off-policy
better
* it doesn't have a "correction" to set the advantage to 0 on the last
term if the state was terminal. I think this correction is an error that
we made (that doesn't matter in practice, since we were never using
terminals; and now aren't using fast_gae).

So, deleting this because it's unused code, and it's not clear that we
should expect to use it in the future.


[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210814958694485)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>