Resample from curricula every time. (#1545)
Use SamplingCurriculum directly instead of wrapping it in
SingleTaskCurriculum, to ensure we re-sample every time we ask for a
task.

Without this change, we end up memoizing configurations. This may be
slightly more efficient, but it works very badly while we still have
randomness in configs. We need to re-sample to ensure that we get
different values for `${sample...}` commands.

This should fix some failures we've been getting where our random maps
aren't actually very random. Or at least not diverse.

Tested by printing when we called the `sample` resolver before and after
this change, and ensured that it made a lot more sense afterward.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210828095522087)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>
Co-authored-by: Jack Heart <jack@stem.ai>