Clean up sampling curricula (#1561)
Does some cleanup to curricula, to make it easier to use the curricula
in-place, instead of always using template files.

* Renames the `env_cfg_template` parameter to `env_cfg_template_path`
for clarity, since sometimes we're working with a config and sometimes
we're working with a file path.
* Replaces separate `bucket_parameters` and `bucket_values` with a
single `sampling_parameters` dictionary -- I don't think there's a
reason for these to be separated.


[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210836080422445)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>