add sparse nav training (#1546)
Add a bucketed sparse navigation task.

The bucketed version adds on to the sparse navigation training we're
doing. It varies map dimensions and available time to complete, as well
as altars. This should allow it to find the best learning regime for big
sparse maps.

I've trained on this a few times with inconsistent results (sometimes it
seems to learn to sweep well; sometimes not). Merging since I think this
is good to build off of.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210828076847588)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>