Heavily reduce the amount of $sampling (#1607)
Remove $sampling from most environment configs by replacing it with the
provided mid value (or an appropriate seeming value when the parameters
were invalid).

The direction we want to move is to have randomness at fewer layers. It
should be controlled above for things we want to be able to adjust in
curricula, and below for things we don't care about. Aka, randomness in
BucketCurricula is okay, and randomness in mapgen is okay, and
randomness in OmegaConf is not.

This change does not preserve any randomness in the files is touches. I
want feedback on that! This doesn't touch the nav curricula, since I
want to preserve randomness there, and so I'll want to convert things to
bucketed curricula; but that takes more work. My hope here is to spend
an amount of effort proportional to how much we care.

So, if any of these are places where we think the randomness is
important, please speak up, and I'll happily do a more random-preserving
change instead.

After this, I'll also need to update $choice.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210853702533283)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>