Fix problematic curricula (#1622)
This fixes a few problems with curricula that came when I was
refactoring other code, and this caused the existing curricula testing
code to catch a few things it had been missing.

Issues addressed:
* Deletes env/mettagrid/multiagent/experiments/cylinder_world. This
references cylinder_world_sequence.CylinderWorldSequence, which appears
to have never existed. I tried to see if cylinder_world.CylinderWorld
was maybe the intended match, but we're trying to pass it `objects`,
which CylinderWorld doesn't take. I think maybe CylinderWorldSequence
was supposed to be checked in as part of
https://github.com/Metta-AI/metta/pull/776, but wasn't.
* Fixes the bucketing keys used in the all_objects curriculum. Before
this fix, the keys didn't line up with the properties I'm pretty certain
they were meant to impact. I _think_ this means we would use the values
in the training template, rather than using the values provided by the
bucket.
* Replaces the default values in the easy_all_objects.yaml with `???` to
ensure we don't accidentally use the defaults. I.e., makes sure that we
know we're getting the bucket values.
* Exclude easy_all_objects.yaml from our env validation test, since now
it's got these ??? that need to be provided. It's a hard knock life.

Tested all this by running `metta test tests/test_validate_all*`

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210864436487760)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>