Allow lots more agent_ids (#1623)
This increase the data type for agent_ids from 8 bit to 32 bit, allowing
us to have way more agents than we can possibly afford.

I don't know of a reason that agent_id would need to be limited to 8
bits; though having it be outside of the byte range does preclude us
from using it as an observation (something we don't currently do, but
might want to consider in the future).

Did a basic training smoke test, but not with > 256 agents.


https://app.asana.com/1/1209016784099267/project/1209096222434381/task/1210852075753523

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210864886182502)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>