allow inline bucketed configs (#1615)
Allows BucketedCurriculum to accept a DictConfig directly

This delays full instantiation of our config, which allows us to inline
configs and still update the config with buckets / overrides.

The code feels convoluted enough (e.g., around the timing of
instantiation) that I'm a 3/5 nervous about the change, but I've added a
legit-feeling test, and the code is protected by tests in general.

Fwiw, I hadn't planned to write this, but I found that the configs I was
writing so obviously wanted to be inline bucket configs that I ended up
here anyways.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210863470885721)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>