Clean up movement tests (#1680)
Cleans up movement tests in a few ways:
* Removes a test that ensures the movement metrics are reset when we
reset the environment. Given that we entirely throw away the cpp
environment when we reset, the test seems superfluous. It's possible
we'll change this reset expectation in the future, but I don't think we
need preemptive tests around that.
* Changes naming on tests to not talk about "refactored".
* Moves a test to the right directory.
* Stops pre-creating levels, and instead fills out the OmegaConf to make
a random level generator. This is the actual main part of the PR! We use
the ability to pass `level` to MettaGrid only in tests, and barely even
there. I think we should remove this functionality in the name of
simplicity.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210887013269347)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>