Fix keys in all_objects curriculum (#1645)
I was exploring changing our bucketing to only allow overwriting
existing keys. This ended up seeming too limited, since we have a lot of
sort-of-implicit keys like "agent.reward.resources.red_ore". This seems
like a fine key for folks to bucket, even if the key isn't in the
explicit configuration -- the error checking happens a little downstream
in the Pydantic GameConfig object.

But in this exploration, I ran into `all_objects.yaml`, which seemed to
just have an invalid config, since it was overriding "initial_items",
which has been switched to "initial_resource_count". So this fixes that.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210877044770751)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>