Pass skypilot secrets with env (#1791)
Previously, we mounted `.netrc` and `.metta/observatory_tokens.yaml`,
but mounts don't work correctly with skypilot service accounts.

So I had to pass secrets through env, and then restore config files in
job container.

One downside of this approach is that we're leaking secret values in
skypilot dahsboard:

![Screenshot 2025-07-29 at
8.26.55 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/evCtXUW5JNyNfqWDHydW/4fe23d24-9ac3-4b70-a6d3-4fb5fd60116a.png)

But that's not a big deal, there were ways to steal these even before.
And there's some benefit in not sending the entire `.netrc` file to
skypilot, because it could contain other credentials.

The combination of this PR and #1803 (which is already deployed) will
fix the service accounts.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210922922355367)