Remove SamplingTaskCurriculum and ProgressiveTaskCurriculum (#1838)
This started as "remove SamplingCurriculum since it doesn't add value
over SingleTaskCurriculum"; but the only user of SamplingCurriculum is
ProgressiveTaskCurriculum, and we can kill that too.

ProgressiveTaskCurriculum is a neat toy example, but it only scales
size, and isn't used, so I'm removing vs refactoring it.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210933452783445)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>