Remove ProgressiveMultiTaskCurriculum (#1871)
Remove the Progressive Curriculum implementation and related tests.

In conversation with @jacklionheart , it sounds like this doesn't fit
the model of where we're taking curricula. Given that this appears to be
basically-not-in-use, it seems like we should delete it. But consider
this PR to be very much a proposal vs a "we're going to do this". So
interested in checking in with @mstormbull about whether this seems
right or not.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210943617511268)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>