Explicitly call `torch.cuda.set_device(device)` (#1950)
This is some preparation for #1882.

There is some hidden and highly implicit piece of state in PyTorch that
keeps track of the device on which a process should run its
calculations. This can be set using `torch.cuda.set_device(device)`
which we have never done so far, as this also implicitly happens during
DDP wrapping of the policy. If not explicitly set, the cuda_device is
initialized to be the same on all local ranks. This has so far never
been a problem as we never attempted to do any collective operations
before the wrapping of our policy. My previous PR changed that as we had
to broadcast some metadata for our PolicyStore before the policy could
be wrapped.

The simple fix for this problem is to explicitly call
`torch.cuda.set_device(device)` as early as possible.

**Testing**
Ran training on 2 nodes with 4 GPUs each (skypilot job 4085).

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210990213301431)