Experience buffer and memory state refactors (#1767)
**Changes**
1. Experience buffer is now flexible. You can assign it tensors to store
via losses' `get_loss_experience_spec` function and the agent's
`get_agent_experience_spec` method. Its primary objective is to support
losses and, secondarily, agent replay.
2. Recursive memory (i.e. state) is now fully encapsulated by the layer
in agent that needs it. Agents can therefore readily swap the LSTM layer
for other types of state-full layers, have multiple memory layers, or
none at all. The rest of the codebase is agnostic to your memory
architecture.
3. The agent (MettaAgent or otherwise) now receives a TensorDict in its
forward and is expected to return one with certain keys such as
"entropy". In addition to simplifying the code, it also affords losses
access to all hidden activations written to the agent's internal
TensorDict. Experience takes what you tell it to and dumps the rest.
4. LSTM state now resets on environment dones and truncateds during
rollout. This happens regardless of BPTT or desync settings.
5. These changes, taken together, are a cleaner way to handle experience
and state so I mopped up a lot of code.

This PR is not backwards compatible with previously trained agents.

**Issues and remaining work:**
1. SPS is about 5% worse than before. I believe this is due to the
extensive use of TensorDict in the agent's forward, particularly the
functions forward_inference and forward_train. I think it's worth the
trade because of the TensorDict's ability to make all relevant tensors
computed in the agent available outside of the agent without needing to
specify which tensors we care about.
2. The wrapper for non-MettaAgent policies needs to take state
management into account.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1210983471897819)