Fix to support torch compile (#2009)
Adjusted how non-tensor information is stored in the tensordict. Code is
longer but it's what's needed to support compile. Note that I tried
NonTensorData but compile caused it to hang - no error message. Switched
back to viewing the data as a tensor with the appropriate size. Requires
more lines of code but works.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211010867626691)