'make test-setup' for 'metta install'. pytests (#2007)
pytests for the metta setup process.

also a few fixes in setup components to make it happen.

this PR does change a few setup components. it modifies them with checks
to environment variables to see if we're in a mode where we shouldn't do
interactive things, like wait for user input at the terminal or open the
user's browser for input. this is so Docker containers and GitHub
Actions workflows can install and test our code properly. these setup
tests are a good example because they'll run in GitHub actions.

i added to `.github/workflows/checks.yml` to get install environments
working as a github action.

`make test-setup` to test locally

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211008375122742)

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>