debug skypilot fails on curricula with diverse levels (#2070)
Currently skypilot jobs training on arena/random and
arena/learning_progress curricula are broken. This is because _level was
not being reset properly in MettaGridEnv. This is a patch for that so
that curricula with different map configs can be trained on in the
short-term, noting that this is not a future-proof solution.