Adding temporary box objects (#2089)
Adding a box object that inherits from HasInventory. The box can be
placed on an empty square faced by an agent by using the "place_box"
action. If boxes are enabled in the environment, the default is for each
agent to be able to only place one box. The "get_output" action used on
a box picks up a box and returns it to the agent that placed it.

You can enable the box object by adding box to game.objects (e.g.)
box:
   type_id: 12
      resources_to_create:
        battery_blue: 1
        


[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211048735304634)

---------

Co-authored-by: Gregory Pylypovych <gregorypylypovych@Mac.localdomain>
Co-authored-by: Gregory Pylypovych <gregorypylypovych@Gregorys-MacBook-Pro-2.local>
Co-authored-by: Richard Higgins <richard@relh.net>