Add leaderboards to observatory (#2078)
The PR has 3 components:

1) Leaderboard definition: essentially like a standard scorecard, but
with a start date instead of specific selected policies / training runs.
The user must specify the set of evals, the metric, and the start date,
and the system maintains the leaderboard .

2) Leaderboard updater: Runs periodically, and for all live dashboards
computes the (leaderboard_id, policy_id -> score) mapping in a table.

3) Leaderboard scoreboard route: uses the leaderboard_policy_scores
table to construct a standard scorecard

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211039460317057)

Co-authored-by: Paul Tsier <paul@elementanalytics.com>