Remove docs/ (#2262)
Reviewing the content here, it seems to be a bunch of random things that
would be much better documented in-context with their code.

I want to be able to build (llm codegen) tooling around the assumption
that in-context README.md docs are our primary form of documentation, so
let's do it that way.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211107842107144)

Co-authored-by: Jack Heart <jack@softmax.com>