Post-dehydration quenching: fix eval workers (#2299)
This fixes two post-dehydration issues with eval workers:
1. we were calling the wrong func
2. change our serialization of the list of simulations to dodge
Omegaconf half-parsing. A possibly better path would be to write the
simulations to a local temp file and point to that. We can switch to it
after Slava implements that more broadly for run.py args

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211127647408336)