remove stale yaml files (#2357)
I believe neither of these should be in used anymore. If they are in
use, it should be in tests. So let's find out!

Tested by grepping for usage, but "benchmark" is a generic enough term
that I'm not sure my grep was successful.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211166645669051)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>