Reduce the number of static_casts we're using (#2407)
We have static_casts in a lot (a few?) of places where we shouldn't need
them. Notably, we should be comfortable with implicit casts when we're
implicitly casing to a wider type, and we should reserve static_casts
for when we're doing something that could potentially lose data. This
will preserve the effect of a static_cast being somewhat of a warning
sign when people see it.

This removes some unneeded static_casts. These came up when we were
casting integral types to floats, and casting InventoryAmounts to
InventoryDeltas, both of which are widening (though int -> float can
certainly be lossy).

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211187562980032)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>