Switched place_box to use the normal resources_consumed (#2411)
place_box previously reimplemented the way that resources are consumed
for an action. This brings it back into the fold, using
`resources_consumed` on the `place_box` action.

Aside from that, this PR intends to be a noop.

We currently have no maps that use this functionality, and I haven't
tested it aside from ensuring it compiles.

I intend to follow up with
(a) making the box be created and destroyed, rather than magically
teleporting to a null square
(b) renaming this to "flag", since the agent is clearly hiding a flag,
not a box

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211188424313258)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>
Co-authored-by: Richard Higgins <richard@relh.net>