Remove no_agent_interference (#2412)
My understanding is that we added this because a good strategy in
capture-the-box is to just body-block your opponent to stop them from
getting your box.

But we already have solutions for that. We can just let the finding
agent freeze the hiding agent for a couple of ticks, and then the
finding agent can swap with the hider.

I could also imagine that we want to make sure that the finder can't
just follow the hider to see where the flag is hidden. In this case, we
should similarly prefer to have the hider start frozen or something,
rather than to have agents be ghosts.

We could also want to consider allowing more than one object per space
on the grid. In this case, we'll need to reconsider how to do
observations to be compatible with this, vs making things invisible.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211188433029776)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>
Co-authored-by: Richard Higgins <richard@relh.net>