inventory_item_names -> resource_names (#2410)
This [mostly] renames inventory_items to resources with Python code.
Inventory_items was the original name; resources is the new, preferred
name that we've partially switched to. This is a push towards settling
on resources.

This doesn't change things associated with cpp, which I'll switch next.
It also doesn't touch stuff associated with replays, which would require
more consideration for backwards compatibility.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211188286978056)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>