remove unneeded function (#2408)
I'm interested in renaming "inventory" to "resources" -- something that
we're partially through at the moment.

As part of this, I ran into `inventory_format`. I wasn't sure if we
should rename this to "resource_format" or what; but it's really just
"flatten dictionary to list", which seems pretty reasonable to inline.

The only functional change this should be making is that we're now
flattening to a list of tuples instead of a list of lists -- but I
expect that any serialization we're doing will result in this becoming a
list of lists anyways.

Tested by running in play mode locally.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211187782708488)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>