Composable Losses (#2019)
### Goals
Shift from where our codebase was, hardcoded for 1) LSTM memory, 2) PPO
and 3) expecting that we only need logits, value, and entropy for all
computations, to an RL platform affording rapid implimentation of
highly-varied losses and architectures.

Refactoring the RL trainer without a presupposed agent architecture or
loss algorithm has plenty of product design choices and challenges. All
we know is that we have some parameterized model(s) that interacts with
the environment and then wishes to be optimized based on some unknown
loss.

### Strategy
Let each loss be a simulacra of trainer.py. We afford the researcher all
the tools they need to write exotic losses in a single file. For
instance, PPO is contained in a single class in a single file, allowing
one to make a new version of that one file and tweak it however they
wish. We therefore contain iterative, exploratory changes. This also
means allowing losses to get messy in the search for RL breakthroughs
while keeping the rest of the code clean and stable.

The result is lightweight but should allow the researcher to implement a
great many configurations and combine them with others by simply writing
a class or two and updating the agent's config.

### Issues 
- There's some awkwardness with control flow. Trainer is the de facto
driver but curriculum might eventually be a better and more direct
scheduler. I had to create a TrainerState class to pass necessary
information to losses.
- Configs are clunky. I'm expecting major shifts in how we manage cfgs
so I dumped everything loss-related into the agent cfg. This will need
to move.
- Imminent to dos: eliminate Param Layer and replace with independent
losses, add a way to save and reload loss state, add the ability to
change losses attached to the policy when restarting training, update
MettaAgent to gather loss specs from each layer.
- Future work could involve feeding agent stats into BaseLoss.

### Changes
The BaseLoss interface provides all the tools you need to build your
losses. Subclassing it affords the ability to:
- control schedule including operating on different phases (e.g. Phasic
Policy Gradient)
- spec what goes in the replay buffer
- take and use outputs from other losses ran before it
- reuse computations run on the agent by other losses and call different
leaf nodes of the network
- load policies and hold its own policies (e.g. kickstarting,
EMA/momentum/BYOL, First Explore, Then Exploit, etc.)
- spec which losses should be tracked and plotted
- block gradients both up and down
- change optimizer
- generally hack anything else you might need
- Updated LossTracker (renamed Loss) to be programmatically defined.

Outside of losses, the codebase no longer has the priors of PPO and
LSTMs.

### Considerations
With composition, scheduling is a primary consideration. There are two
senses:
1. At which agent_step count do our losses run rollouts or train? This
is tentatively handled by the scheduler in the config but it should
eventually be handled by curriculum and, eventually, the gardener. For
now, each loss class determines should_run based on agent_step. A loss's
schedule in rollout can be different than in train.
2. During a given rollout or minibatch update, which losses run first?
Losses are run in the order they are listed in the config.
This also means that they can share data or process the same data which
provides opportunities for efficiencies. In training, your first loss
samples experience and writes it to the shared loss minibatch TensorDict
for other losses to use. This first loss should also include the policy
output so that subsequent losses don't re-run parts of the network that
already computed. If subsequent losses don't want to use that minibatch
then they're free to instead sample their own and/or to re-run the
network.

Using MettaAgent is slightly different. With enhanced neural
architecture flexility we need to delegate the management of the agent's
state to its layers. Now, we have methods to tell the agent the nature
of our ask. Specifically, this means calling methods that tell it more
about the type of computation we're trying to perform. There are
currently four options: on_new_training_run, on_rollout_start,
on_train_mb_start, and on_eval_start. This is hopefully the simplest set
that is expansive enough to afford the various experiments we might want
to run. As an example, on_eval_start includes things like setting
policy.eval() and reset_memory(). on_rollout_start() could mean reset
memory for certain layers or to store context in a transformer. Layers
therefore have the flexibility to handle a lot of complexity that would
have otherwise muddied the code external to the agent, requiring it
handle various architectures.
When working with layers that require some state, use the layer to
manage its own state if that state should be used in inference or eval.
But if it's a concept that needs to be used across from rollout to
training then use the experience buffer spec.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211061517827638)

---------

Co-authored-by: Richard Higgins <richard@relh.net>
Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>
Co-authored-by: Richard Higgins <richard@softmax.com>