Fix: `metta` binary correctly uses enclosing uv project (#2477)
`./install.sh` sets up a symlink to ~/.local/bin such that `metta` can
be used in lieu of `uv run metta`. to support multiple checkouts, the
symlnik points to a binary that tries to identify the appropriate
checkout and use its uv environment

Our logic for finding the appropriate checkout was previously broken,
which made running `metta` always refer to the `metta_cli.py` in the
checkout from which the symlink was established even if you were
currently within another checkout. this fixes that



[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211230681938882)