collapse some observation helpers (#2519)
In order to make tests robust to changing feature_ids, we'll need to
read the feature ids as they're exposed by the environment. This means
our helpers should be written so we can more consistently pass in
feature_ids. This PR does this by unwinding some of the higher level
helped functions we had. As a sort-of bonus this reduces the number of
helper functions we have significantly. As a penalty, this means our
tests are slightly messier.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211253625530679)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>