Simplify install.sh: - Only supports one form of specifying a target profile - Configures profile and installs in one command, not two; complexity therein moved into metta install - Docs updated
metta install is updated, and gives more explicit error messages about missing profiles. A profile must exist before installing. If installing in non-interactive mode, the target profile must be specified with --profile. If in interactive mode and no profile is specified, the setup wizard will be run.

Quietens metta symlink-setup in the happy case where an existing symlink is found