Simplify install.sh, make `metta install` handle configuring profile more robustly (#2586)
1. We had partial support for `metta install` also taking on the job of
`metta configure` before, but it only ran if a profile didn't already
exist. Now, as a first step, `metta install` does what it can to ensure
a profile gets configured if possible. It configures profile if a
`--profile` is provided, else if a profile already exists on the device
it continues, else if in non-interactive mode it errors, else it runs
the interactive profile configure wizard.

2. `./install.sh` is simplified:
- defers to `metta install` on configuration logic
- only supports one way to specify a profile
- doesnt swallow error logs anymore


[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211337281597637)