make installing pnpm less fragile (#2593)
previously, we were using `pnpm --version` to check if pnpm was
availble, which automatically triggered an interactive prompt. if you
did not interact with that prompt, metta install would hang. if you did,
you would install pnpm through a means we don't intend. now, we use
`shutil.which` to check if pnpm is available, and so end up using
`corepack enable` to install it (as intended).

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211339124994419)