metta install should force install to remediate authentication issues selectively (#2596)
When a component is installed but not connected (authenticated) in the
expected way, we advise a --force install of that component. This PR
updates metta install to only do so in cases where a force install could
possibly reauthenticate the user.

Also actually skip trying to auth with wandb when we say we will

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211339097194013)