[MettaScope2] improve zooming, smooth when in and out (#2606)
related: https://github.com/treeform/windy/pull/127 try to move
platform-specific scrolling stuff to windy instead of use when defined()

scrolling was not smooth, it was very rapid when zoomed out and very
slow when zoomed in.

changed it to be similar to the scaling we had on the typescript
version, much more smooth out to in.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211349147120256)