Add Usable, and make Move trigger Use (#2656)
As per title, this adds an abstract Usable class. This also changes Move
to be (implicitly -- not in name) MoveOrUse. The latter seemed more
reasonable than adding a separate action, since there's not a clear
reason to want both Move and MoveOrUse.

Usable isn't helpful in this PR, since nothing implements it; but I'm
shipping this first to simplify the code from which I'm trying to fix
segfaults.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211386581427044)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>