Adds an assembler object (#2669)
This adds Recipes, and Assemblers which can expose different recipes
depending on how their surrounded by agents. Assemblers keep a vector of
256 recipes, which may be repeats. This keeps the cpp logic of which
recipe to use simple.

Right now resources are consumed in a consistent but arbitrary order.

This includes a translation of Arena to use Assemblers instead of
Converters, and Assemblers appear to work in this context. I haven't
produced more thorough unit tests of this, but will do so reactively as
it becomes more clear what sort of testing we need.


https://app.asana.com/1/1209016784099267/project/1211340635624763/task/1211355235230371

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211390250279726)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>