LSTM that resets on dones and truncates during training (#2793)
Wrapped a policy around the unrolled LSTM layer to make it easier for
experimenters to use. Simply use FastLSTMResetConfig() as your policy.
Import it with from metta.agent.policies.fast_lstm_reset import
FastLSTMResetConfig.

The LSTMReset component in that policy adds functionality to

reset on dones and truncateds during training (we already reset on dones
and truncateds during inference)
start off BPTT segments using the cell state from the previous time step
generated during rollout. This is better because it doesn't start
segments with an empty cell state but also bad because the cell state
goes stale with each minibatch back prop update.
It also uses burn-in and the number of burn in steps are set by the
policy constructor as an attribute of the policy. This is temporary! In
the future, we want burn-in steps attribute to move to be a part of a
rollout spec.

The big downside is that it runs at roughly 60% the speed of the normal
agent due to the fact that the LSTM is "unrolled" during training so it
can check whether it needs to reset dones and truncateds.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211434255695276)

---------

Co-authored-by: Richard Higgins <richard@relh.net>
Co-authored-by: Richard Higgins <richard@softmax.com>