Remove a global observation about what resources give reward (#2794)
Removing this to make it cleaner to switch rewards to be stats based. We
could still preserve this if we felt super passionate about it, but it's
not used at the moment. So my proposal is to get rid of it, and we can
discuss.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211434192054658)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>