Unordered chains (#2780)
Adds unordered chains and an `icl_resouce_chain.py` with superclasses to
help consolidate the chain logic and learning progress parameters.

`unordered_chain.py` is a recipe for flat recipe creation where the task
is to figure out the recipe given a bunch of resources to play with. The
inventory for all objects is capped at one to help ensure interesting
leraning.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211448236948614)

---------

Co-authored-by: jacke <jacke@Jonathans-MacBook-Pro.local>
Co-authored-by: Daphne Demekas <daphnedemekas@Daphnes-MBP-2.localdomain>
Co-authored-by: jacke <jacke@Mac.localdomain>
Co-authored-by: Daphne Demekas <daphnedemekas@Daphnes-MacBook-Pro-2.local>
Co-authored-by: Daphne <45669133+daphnedemekas@users.noreply.github.com>