Added stateful policy (#2879)
Added `stateful_policy.py` and set it as the default in `main.py`. 

In StatefulPolicy, detached state in case we use it in training someday.
Also put state management at the policy level so it can eventually be
updated to index by agent ID if needed in the future.

Deleted an unused decoder layer in SimplePolicy.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211473600471365)