Make assemblers clippable (#2945)
This adds an ability for assemblers to become clipped. This gives them a
second set of recipes that take precedence, and unclips them if one of
those recipes is successful.

This is different than the plan we'd discussed, where we replace the
objects in the grid -- but in our current design, only assemblers need
to be clippable, and this is definitely simpler than the version that
involves replacement.

Right now there's nothing in the game that would cause assemblers to
_become_ clipped, so that's still needed.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211506173640416)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>