Cleans up mettagrid_c_config (#3018)
The main thing this does is remove a number of guards we don't want,
where we silently filter resources that don't exist in the current
environment. Instead, we want to fail.

This also remove checks to filter out when values are non-positive.
Mostly, this removes zero-valued entries. It's at least sort of good to
remove zero-valued entries; but we shouldn't be having zero-valued
entries, so the filtering code feels like sort of low-valued noise.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211526503869765)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>