bring back asan in debug (#2949)
We lost asan when moving from bazel to cmake. Now we bring it back.

Attempted to run msan as well, but it had issues with gtest, and has
been abandoned for now. It was not previously used.

This runs only for `cd packages/mettagrid && make test`, it does NOT run
for any `pytest`.

Tested by putting a memory error in assembler.hpp, which asan caught in
`cd packages/mettagrid && make test`.