cogames train print post-training messages even when users ctrl-c (#3144)
When we finish training, we print instructions to run the trained
policy. Now when we ctrl-c, we do too.

The simplest solution, which is to wrap the training loop in a
KeyboardException try/except, doesn't work because pufferlib binds
SIGINT to exit the process immediately

So we instead support a custom sigint contextmanager that intercepts the
first SIGINT and raises a KeyboardInterrupt (default python behavior).
when the contextmanager exits, it rebinds the old signal handler and
calls it if it received any SIGINTs. if it receives more than one SIGINT
within its context, it triggers that behavior early.

We also warn when a user has cut off training before the first
meaningful checkpoint would have been saved

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211571194637188)