balanced consume_resources_for_recipe (#3237)
This balances resource consumption across participating agents. It
mostly tries to take an even amount from each agent. In cases where
rounding requires some agents to pay more, it happens that agents with
more of that resource will pay more. I also expect the sorting to work
out in a way where it's always the same agents who pay more, if it's
tied. That's not intentional, but hopefully is nbd. In any case, it
seems better than trying to suck-dry the agent in the north west.


https://app.asana.com/1/1209016784099267/project/1209041170403474/task/1211634548177701

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211643572519010)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>