Refactor MapBuilderConfig polymorphic loading (#3216)
After Pydantic 2.12, I was able to do some implementation cleanups here.

The longer-term goal is to come up with the common base
`PolymorphicConfig` class that will be reused in `MapBuilderConfig`,
`SceneConfig` all other config types that need to be serialized with
`type: ClassName` in their YAML/JSON.

But there are a few more roadblocks to that, so I'm proceeding
incrementally.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211628533619685)