Fix silently failing imports (#3336)
This attempts to address (by removing) imports that fail on MacBooks,
and removes logic that silently swallows these import errors.

This stems from me doing work that was causing
`test_recipe_discovery.py` to fail -- specifically, I'd messed up some
imports. However, the code that does the automatic recipe discovery
swallows import errors, so the test failures showed up as "some things
are missing" rather than "you messed up some imports". So the point of
this is is _mostly_ to make it so such errors are more clear in the
future.

As part of removing this import error protect, I found a handful of
places where we appear to be importing things that don't exist. I don't
know if they're supposed to exist in some other context, so input on the
PR is welcome. If these imports _are_ supposed to only exist in some
other, not macbook context, then I'd like us to find some way to make
that clear.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211678392074479)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>