[MettaScope] Vibe action queue (#3400)
-
https://app.asana.com/1/1209016784099267/project/1209513563278644/task/1211724355992402
- add support for vibe actions to the shift (and shift-r) queues
- re-name the destination queue to objective queue, to make it more
generic to goals to do not involve locations (like performing a vibe)

- when testing the vibes work properly in queues, but I am seeing some
issues with agent movement that also reproduce on the main branch. if an
agent reaches it's destination, it continues moving in the same
direction, despite not having a path. I think this might be related to
action reworking?

- action arguments are deprecated. they are used in this PR for vibes
still.
- we can update all of our action logic in the future for
https://app.asana.com/1/1209016784099267/project/1209513563278644/task/1211724355992412
  - this might be related to the movement issues?

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211736363782533)