MCP server for embeddings based PR search (#3404)
## Objective
We want our coding agents to be able to access knowledge about past bugs
and common solutions.

## Approach
We calculate text embeddings for each PR description and give Claude
Code and Codex access to an MCP server that allows a query as input. We
calculate the similarity between the query and each PR description and
the top-n closest PR descriptions.
For PRs that have no description or which have a description that is
shorter than three lines, we prompt an LLM to write a description for us
with the full diff provided as context.

## Details
### Alternative approach
At first I explored the idea of having an LLM take the git history and
write a document explaining common pitfalls in our repo, but that seemed
suboptimal as it was hard to compress the large number of small bugs we
had in any way without losing too much information. Embedding based
search seems to be the better approach given the large number of PRs
that has to be made accessible.
### Technical details
In order to calculate the query embedding, users need a Gemini API key.
This key will be retrieved from the AWS secret store during `metta
install` if the user selected employee as user type.
The PR embeddings are stored in an S3 bucket to avoid having every user
calculate the embeddings. They are downloaded and cached during `metta
install`.
We will later create an AWS Lambda job (or similar) to update the cache
files in some interval.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211736754743590)