restore old mettascope assets in gridworks (#3427)
Previously, gridworks symlinked to mettascope2 png tiles.

But mettascope2 tiles are currently in the flux because of pixel art
migration, and Andre has agreed that copy-pasting for now is a better
option.

Maybe I'll be able to restore the symlink later when things stabilize.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211748358358070)